Move tests to NestedDataSourceSuite
